Change: Move VersionError into and own module
